requirements.txt
Django==1.6.2
South==0.8.4
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-crispy-forms==1.4.0
django-haystack==2.1.0
django-registration==1.0
django-taggit==0.11.2
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.2
pyelasticsearch==0.6.1
pystache==0.5.3
requests==2.2.1
simplejson==3.3.3
six==1.6.1
static==1.0.2
wsgiref==0.1.2
